requirements.txt
asgiref==3.4.1
attrs==22.2.0
beautifulsoup4==4.12.0
branca==0.5.0
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
cryptography==40.0.2
Django==3.2
django-ckeditor==6.1.0
django-countries==7.5.1
django-debug-toolbar==3.2.4
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-multiselectfield==0.1.12
django-ranged-response==0.2.0
django-resized==1.0.2
django-simple-captcha==0.5.17
Fiona==1.8.22
httplib2==0.22.0
idna==3.4
importlib-metadata==4.8.3
Jinja2==3.0.3
MarkupSafe==2.0.1
munch==4.0.0
mysqlclient==2.1.1
Pillow==8.4.0
# protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pymemcache==3.5.2
pyOpenSSL==23.2.0
pyparsing==3.1.0
# pyproj==3.0.1
python-dateutil==2.8.2
python-dotenv==0.19.1
python-memcached==1.59
pytz==2022.7.1
requests==2.27.1
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.15
qrcode==7.4.2
whitenoise==5.3.0
zipp==3.6.0
django-recaptcha==4.0.0
# reportlab==4.0.6

# folium
# earthengine-api
# pandas
# geopandas
# django-payments